Fix typechecking in worker.
Previously, the right way to indicate that a module should be checked as
a web worker was not clearly documented. Now it is:
  https://deno.land/manual@v1.11.2/typescript/types

This breaks if types are imported from the worker module though, so
split the exported types to a separate worker_api module.
